Made Controller for views
I made 3 controllers for museums, operators and evidences table views.
They are all in a file mainCtrl.js
So it's more readable. Then in the respective files I just added <script
src='mainCtrl.js'></script> .
About getting json from a page and not from a file, i dont know if it
works, but i think we cannot try to use local server because eclipse
uses a special version of tomcat which is integrated in eclipse, so
maybe it's not the same localhost and thats why doesnt work.