Some new change in Backoffice
Added ng-model in registration  files. So updated these files. I created
a prototype of the page which will show the single evidence (it's a copy
of the registration page but readonly fields. I cant put the value from
json to the field yet.
TODO: update show-evidence so it will shows the actual values of the
evidence chosen.